UPD: Move global chat outside
