+++ src/pages/global-chat/styled.js
